:hammer:[rename]: @components/index 경로 통합
