+++ src/components/home/index.ts
@@ -3,3 +3,4 @@ export { default as DayCell } from './DayCell';
 export { default as DateSummary } from './DateSummary';
 export { default as TotalSummary } from './TotalSummary';
 export { default as TransactionItem } from './TransactionItem';
+export { default as PlusBtn } from './PlusBtn';
