+++ src/components/index.ts
@@ -10,3 +10,5 @@ export { default as ProfileInfo } from './ProfileInfo';
 export { default as StatusTag } from './StatusTag';
 export { default as CategoryItem } from './CategoryItem';
 export { default as StatsSummary } from './StatsSummary';
+export { default as CustomDatePicker } from './CustomDatePicker';
+export { default as ScrollToTop } from './ScrollToTop';
